chore: refactore interfaces and tokens
