+++ projects/ngx-airtable/src/lib/interfaces.ts
@@ -1,17 +1,7 @@
-import {RunActionMethod, SortDirection} from './types';
+import {RunActionMethod} from './types';
 import {Base} from './node-port/base';
 import {HttpHeaders, HttpParams} from '@angular/common/http';
-
-export interface AirtableConfiguration {
-  apiKey?: string;
-  endpointUrl?: string;
-  apiVersion?: number;
-}
-
-export interface TableOptions {
-  tableName?: string;
-  tableId?: string;
-}
+import { TableOptions } from './interfaces/table-options';
 
 export interface RunActionOptions {
   base: Base;
@@ -20,24 +10,6 @@ export interface RunActionOptions {
   params?: HttpParams;
   body?: any;
   headers?: HttpHeaders;
-}
-
-export interface SelectParams {
-  fields?: string[];
-  filterByFormula?: string;
-  maxRecords?: number;
-  pageSize?: number;
-  sort?: SortParam[];
-  view?: string;
-}
-
-export interface Params {
-  [key: string]: any;
-}
-
-export interface SortParam {
-  field: string;
-  direction: SortDirection;
 }
 
 export interface Link {
