+++ projects/ngx-airtable/src/lib/interfaces/airtable-configuration.ts
@@ -1,0 +1,5 @@
+export interface AirtableConfiguration {
+    apiKey: string;
+    endpointUrl?: string;
+    apiVersion?: number;
+}
