+++ projects/ngx-airtable/src/lib/interfaces/executioner.ts
@@ -1,0 +1,12 @@
+import { HttpClient, HttpParams, HttpHeaders } from '@angular/common/http';
+import { ExecutionMethod } from '../types/execution-method';
+
+export interface Executioner {
+    http: HttpClient;
+    base?: string;
+    method?: ExecutionMethod;
+    url: string;
+    httpParams?: HttpParams;
+    body?: any;
+    headers?: HttpHeaders;
+}
