+++ projects/ngx-airtable/src/lib/interfaces/params.ts
@@ -1,0 +1,3 @@
+export interface Params {
+    [key: string]: any;
+}
