+++ projects/ngx-airtable/src/lib/interfaces/sort-param.ts
@@ -1,0 +1,6 @@
+import { SortDirection } from '../types/sort-direction';
+
+export interface SortParam {
+    field: string;
+    direction: SortDirection;
+}
