+++ projects/ngx-airtable/src/lib/interfaces/table-options.ts
@@ -1,0 +1,4 @@
+export interface TableOptions {
+    tableName?: string;
+    tableId?: string;
+}
