--- projects/ngx-airtable/src/lib/token.ts
@@ -1,4 +1,0 @@
-import {InjectionToken} from '@angular/core';
-import {AirtableConfiguration} from './interfaces';
-
-export const AirtableConfigToken: InjectionToken<AirtableConfiguration> = new InjectionToken<AirtableConfiguration>('globalConfiguration');
