+++ projects/ngx-airtable/src/lib/types/cell-format.ts
@@ -1,0 +1,1 @@
+export type CellFormat = 'json' | 'string';
