+++ projects/ngx-airtable/src/lib/types/execution-method.ts
@@ -1,0 +1,1 @@
+export type ExecutionMethod = 'GET' | 'POST' | 'PATCH' | 'PUT' | 'DELETE';
