+++ projects/ngx-airtable/src/lib/types/sort-direction.ts
@@ -1,0 +1,1 @@
+export type SortDirection = 'asc' | 'desc';
